test(code): retire the golden resolver snapshot for named tests
The 2,398-case snapshot proved the ranked resolver reproduced the legacy
engine before any consumer migrated. With the ranked engine authoritative,
the hash-keyed baseline can no longer distinguish a deliberate fix from an
absorbed regression, so the matrix goes away and the named integration tests
carry the deliberate behaviors forward: exit 78 enforcement, fail-closed
provider health, deny-list reader parity, quoted dotted keys, managed-path
writes, and reload/preview snapshot isolation.
